% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{news}
\alias{CytobankBridgeR_news}
\alias{news}
\title{News}
\usage{
CytobankBridgeR_news()
}
\description{
Get news on CytobankR updates
}
\details{
\code{CytobankR_news} View a log of CytobankR updates and release notes.
}

