% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bubble_spade_nodes.R, R/spade.R
\name{spade.bubble_spade_nodes}
\alias{spade}
\alias{spade.bubble_spade_nodes}
\title{SPADE}
\usage{
spade.bubble_spade_nodes(UserSession, spade, prefix = "cluster_")
}
\arguments{
\item{UserSession}{Cytobank UserSession object (created via the \link[CytobankAPI]{authenticate} function)}

\item{spade}{Cytobank \link[CytobankAPI]{spade} object}

\item{prefix}{character representing a prefix to be used before each node number}
}
\description{
SPADE High Level Functions
}
\details{
\code{spade.bubble_spade_nodes} Bubble each individual node on a SPADE tree as their own bubble, with a specific prefix.
}
\examples{
\donttest{# Using the prefix "bubble_" will provide bubbles in the form:
#   "bubble_1", "bubble_2", "bubble_n", etc...
spade.bubble_spade_nodes(cyto_session, spade=cyto_spade, prefix="bubble_")
}
}

